Use intValue() to avoid format conversion error in audit.

+++ intermediaries/subscription-notification-service-intermediary-common/src/main/java/org/ojbc/intermediaries/sn/subscription/SubscriptionProcessor.java
@@ -74,7 +74,7 @@ public abstract class SubscriptionProcessor {
         
         Number subscriptionId = subscriptionSearchQueryDAO.subscribe(request, new LocalDate());
         
-        exchange.getIn().setHeader("subscriptionId", subscriptionId);
+        exchange.getIn().setHeader("subscriptionId", subscriptionId.intValue());
         
         exchange.getMessage().setBody(SubscriptionResponseBuilderUtil.createSubscribeResponse());
     }
